Wrap types in object types and import in generated code
