+++ src/test/resources/apollo/blog/AddArticle.scala
@@ -1,4 +1,5 @@
 import sangria.macros._
+import types._
 object AddArticle {
   object addArticle extends GraphQLQuery {
     val document: sangria.ast.Document = graphql"""mutation addArticle($$content: ArticleContent!) {
@@ -11,7 +12,4 @@ object AddArticle {
     case class Data(addArticle: AddArticle)
     case class AddArticle(id: ID, title: String)
   }
-  case class ArticleAuthor(id: ID)
-  case class ArticleContent(title: String, body: String, tags: Option[List[String]], author: ArticleAuthor)
-  type ID = String
 }
