+++ src/test/resources/apollo/blog/AddArticleTypes.scala
@@ -1,0 +1,5 @@
+object types {
+  case class ArticleAuthor(id: ID)
+  case class ArticleContent(title: String, body: String, tags: Option[List[String]], author: ArticleAuthor)
+  type ID = String
+}
