+++ src/test/resources/apollo/blog/BlogArticleQueryTypes.scala
@@ -1,0 +1,9 @@
+object types {
+  case class ArticleQuery(ids: Option[List[ID]], statuses: Option[List[ArticleStatus]])
+  sealed trait ArticleStatus
+  object ArticleStatus {
+    case object DRAFT extends ArticleStatus
+    case object PUBLISHED extends ArticleStatus
+  }
+  type ID = String
+}
