+++ src/test/resources/apollo/blog/BlogByID.scala
@@ -1,4 +1,5 @@
 import sangria.macros._
+import types._
 object BlogByID {
   object Blog extends GraphQLQuery {
     val document: sangria.ast.Document = graphql"""query Blog($$blogId: ID!) {
@@ -10,5 +11,4 @@ object BlogByID {
     case class Data(blog: Blog)
     case class Blog(title: String)
   }
-  type ID = String
 }
