+++ src/test/resources/apollo/starwars-circe/HeroAndFriends.scala
@@ -1,6 +1,7 @@
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
 import sangria.macros._
+import types._
 object HeroAndFriends {
   object HeroAndFriends extends GraphQLQuery {
     val document: sangria.ast.Document = graphql"""query HeroAndFriends {
