+++ src/test/resources/apollo/starwars-circe/InputVariables.scala
@@ -1,6 +1,7 @@
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
 import sangria.macros._
+import types._
 object InputVariables {
   object InputVariables extends GraphQLQuery {
     val document: sangria.ast.Document = graphql"""query InputVariables($$humanId: String!) {
