+++ src/test/resources/apollo/starwars-circe/SearchQuery.scala
@@ -1,6 +1,7 @@
 import io.circe.Decoder
 import io.circe.generic.semiauto.deriveDecoder
 import sangria.macros._
+import types._
 object SearchQuery {
   object SearchQuery extends GraphQLQuery {
     val document: sangria.ast.Document = graphql"""query SearchQuery($$text: String!) {
