+++ src/test/resources/apollo/starwars/EpisodeEnumTypes.scala
@@ -1,6 +1,8 @@
-sealed trait Episode
-object Episode {
-  case object NEWHOPE extends Episode
-  case object EMPIRE extends Episode
-  case object JEDI extends Episode
+object types {
+  sealed trait Episode
+  object Episode {
+    case object NEWHOPE extends Episode
+    case object EMPIRE extends Episode
+    case object JEDI extends Episode
+  }
 }
