+++ src/test/resources/apollo/starwars/HeroFragmentQuery.scala
@@ -1,4 +1,5 @@
 import sangria.macros._
+import types._
 object HeroFragmentQuery {
   object HeroFragmentQuery extends GraphQLQuery {
     val document: sangria.ast.Document = graphql"""query HeroFragmentQuery {
