+++ src/test/resources/apollo/starwars/InputVariables.scala
@@ -1,4 +1,5 @@
 import sangria.macros._
+import types._
 object InputVariables {
   object InputVariables extends GraphQLQuery {
     val document: sangria.ast.Document = graphql"""query InputVariables($$humanId: String!) {
